feat: stabilize runtime ownership gate
