+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/TelemetryService.kt
@@ -53,9 +53,8 @@ class TelemetryService(
   }
 
   fun autoSync(dbOverride: String? = null) {
-    val settings = telemetrySettingsOrNull(settingsProvider) ?: return
-    if (!settings.enabled) return
-    if (!database.databaseExists(dbOverride)) return
+    val settings = telemetrySettingsOrNull(settingsProvider)
+    if (settings == null || !settings.enabled || !database.databaseExists(dbOverride)) return
     database.transaction(dbOverride) { unitOfWork ->
       TelemetrySyncRuntime.autoSyncTelemetry(settings, unitOfWork.telemetryOutbox, telemetryClient)
     }
